made_simluation/requirements.txt
addict==2.4.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
cachetools==4.2.0
chardet==4.0.0
cycler==0.10.0
cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
easydict==1.9
entrypoints==0.3
filterpy==1.4.5
fire==0.3.1
glob2==0.6
idna==2.10
imagecorruptions==1.1.2
imageio==2.9.0
importlib-metadata==3.4.0
ipdb==0.13.7
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lap==0.4.0
llvmlite==0.36.0
markupsafe==1.1.1
matplotlib==2.2.3
mistune==0.8.4
mmcv==1.2.5
motmetrics==1.1.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5.1
notebook==6.2.0
numba==0.53.1
numpy==1.21.2
nuscenes-devkit==1.0.5
objgraph==3.5.0
opencv-contrib-python==4.5.2.54
opencv-python==4.5.3.56
packaging==20.8
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.3.1
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.17.3
ptyprocess==0.7.0
pybind11==2.6.2
pycocotools==2.0.2
pygments==2.7.4
pyparsing==2.4.7
pyquaternion==0.9.9
pyrsistent==0.17.3
pytest-runner==5.3.1
python-dateutil==2.8.1
# python-graphviz==0.16
pytz==2021.1
pywavelets==1.1.1
pyyaml==5.3.1
pyzmq==21.0.1
qtconsole==5.0.1
qtpy==1.9.0
requests==2.25.1
scikit-image==0.18.2
scikit-learn==0.19.2
scipy==1.6.0
seaborn==0.11.1
send2trash==1.5.0
setuptools==57.4.0
shapely==1.7.1
# spconv==1.2.1
tensorboardx==2.4
termcolor==1.1.0
terminado==0.9.2
terminaltables==3.1.0
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.7.2
toml==0.10.2
torchgeometry==0.1.2
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yapf==0.30.0
zipp==3.4.0